add rendered points state
