+++ components/Map/map.tsx
@@ -1,8 +1,12 @@
+/* eslint-disable semi */
+/* eslint-disable prettier/prettier */
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/* eslint-disable quotes */
 import React, { useEffect, useState } from "react";
 import { StyleSheet, View, PermissionsAndroid, Platform } from "react-native";
 
 import Geolocation from "react-native-geolocation-service";
-import Toast from "react-native-root-toast";
+
 import Mapbox from "@rnmapbox/maps";
 import { getPointAnnotation, getLineAnnotation } from "../../services";
 
@@ -50,9 +54,7 @@ const requestLocationPermission = async () => {
 
 const Map = ({ navigation }: any) => {
   const [currentLocation, setCurrentLocation] = useState([0, 0]); // Longitude, Latitude
-
-  //   let startingPoint = [-6.2653554, 53.324153];
-  let destinationPoint = [-6.2650513, 53.3256942];
+  const [renderedPoints, setRenderedPoints] = useState<any>([])
 
   useEffect(() => {
     if (Platform.OS === "android") {
@@ -82,37 +84,7 @@ const Map = ({ navigation }: any) => {
     setCurrentLocation([0, 0]);
   };
 
-  const getLocationData = async (data: any) => {
-    try {
-      // const response = await fetch('https://reactnative.dev/movies.json');
-
-      const response = await fetch(
-        `http://ecocompass.anupal.me/test-core?latitude=${encodeURIComponent(
-          data.latitude
-        )}&longitude=${encodeURIComponent(data.longitude)}`,
-        { method: "GET" }
-      );
-
-      const json = await response.json();
-      Toast.show(json.message, {
-        duration: 10000,
-      });
-      console.log(json);
-    } catch (error) {
-      // console.error(error);
-    } finally {
-      // setLoading(false);
-    }
-  };
-
-  const getUserClickLoc = function (loc) {
-    let locObj = loc.nativeEvent.coordinate;
-    getLocationData({
-      latitude: locObj.latitude,
-      longitude: locObj.longitude,
-    });
-  };
-
+  // sample route
   const route: any = {
     type: "FeatureCollection",
     features: [
@@ -139,21 +111,36 @@ const Map = ({ navigation }: any) => {
     ],
   };
 
+  const userLocationUpdate = (data: any) => {
+  };
+
+  const getClickedPoint = (feature: any) => {
+    setRenderedPoints([getPointAnnotation({id: 'abc', coordinates: feature.geometry.coordinates})])
+  }
+
+  const pointsArr = (coords: any, id: any) => {
+    getPointAnnotation({
+      coordinates: coords,
+      id: id,
+    })
+  }
+
   return (
     <View style={styles.page}>
       <View style={styles.container}>
-        <Mapbox.MapView style={styles.map}>
+        <Mapbox.MapView
+          style={styles.map}
+          onPress={getClickedPoint}
+        >
           <Mapbox.Camera
             zoomLevel={14}
             centerCoordinate={currentLocation}
             animationMode={"flyTo"}
             animationDuration={2000}
           />
-          {getPointAnnotation({
-            coordinates: currentLocation,
-            id: "currentLocation",
-          })}
-          {route && getLineAnnotation({ route })}
+          {renderedPoints}
+          <Mapbox.UserLocation onUpdate={userLocationUpdate} />
+          {/* {route && getLineAnnotation({ route })} */}
         </Mapbox.MapView>
       </View>
     </View>
