+++ services/network.service.ts
@@ -1,0 +1,32 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/* eslint-disable prettier/prettier */
+import Toast from "react-native-root-toast";
+
+const getLocationData = async (data: any) => {
+    try {
+        // const response = await fetch('https://reactnative.dev/movies.json');
+
+        const response = await fetch(
+            `http://ecocompass.anupal.me/test-core?latitude=${encodeURIComponent(
+                data.latitude
+            )}&longitude=${encodeURIComponent(data.longitude)}`,
+            { method: "GET" }
+        );
+
+        const json = await response.json();
+        Toast.show(json.message, {
+            duration: 10000,
+        });
+        console.log(json);
+    } catch (error) {
+    } finally {
+    }
+};
+
+const getUserClickLoc = function (loc: any) {
+    let locObj = loc.nativeEvent.coordinate;
+    getLocationData({
+        latitude: locObj.latitude,
+        longitude: locObj.longitude,
+    });
+};
